requirements.txt
streamlit
google-generativeai
python-dotenv
pillow
pypdf2
pdf2image